Reorg, comments and metrics
